Remove `--locked-schema` when running migrations on Heroku
This reverts a portion of #1611.  Heroku does not make a checkout of
the source code available at runtime, so at the point in time we run
migrations `src/schema.rs` will not exist, and will thus always fail
this check.

The `--locked-schema` check is still run on CI, so this still gives us
confidence that our schema and `src/schema.rs` do not diverge.