docker/frontend: Correctly proxy to backend
When using `npm run start --proxy http://backend:8888`, `npm` interprets
the `--proxy http://backend:8888` as a parameter that IT should parse,
which means the parameter never gets to `ember serve`. Add a `--` in the
middle so that `npm` knows not to parse that flag and instead pass it to
`ember serve`.