Properly downcast `swirl::EnqueueError`
The return type of `enqueue` is `Box<dyn Error + Send + Sync>`.
Unfortunately, `Box<dyn Error>` does not implement `Error`, so our from
impl doesn't apply. We're also not able to add an explicit `From` impl
either (which is odd, coherence rules should let us write that impl).

I've opted to just add an explicit converstion function that we can use
here. We should probably check that there's nowhere else we're passing
an error object to `human` or `internal`, but that will come in a
separate PR