Change Travis to PG 11
We're running PG 11 in production now, so we should test against the
same version in CI. I had to put us on the newer xenial image to get
this to work, since trusty leaves 9.2 running if it doesn't recognize
the newer version you're trying to run, which caused issues.

There's no built-in cookbook for 11 yet, so I've had to install it from
apt and copy some additional minimal configuration to get things to
work.