Fix new instances caught by `clippy::redundant_closure`
The lint is more sensitive starting in 1.34 (currently in beta).  This
addresses the new warnings so that CI isn't affected after the release
next week.

There is one case that recommends using `failure::error::Error::compat`
however we do not currently include `failure` in our TOML file.  Rather
than add it just for this case, I've allowed the lint for this
expression.