Fingerprinting SVGs
In the production build all the SVGs are copied into dist/assets without fingerprints. All other assets have fingerprints. Since all files in dist/assets are served with far-future cache headers there would be cache issues if the files were ever updated. This is not major issue as these files are not used in normal operations.

Closes #1575