Remove duplicate calls to `db_conn` in a single endpoint
These endpoints are obtaining two separate connections for each
request.  If there is a backlog of requests, then these endpoints will
wait in line multiple times for a database connection.

Any request with an auth token or cookie will also obtain and drop a
connection in the `CurrentUser` middleware.  We could consider caching
and reusing this database connection as well.