Don't allow crate names in use to be added to `reserved_crate_names`
We've made this mistake twice now. If a record gets added to this table
and a crate already exists with that name, we will fail to update
download statistics as we can no longer update that row in `crates`. The
expected invariant is that the crates table never overlaps with
`reserved_crate_names`, we should enforce that remains true.