allow configuring the application's domain name
This commit allows setting the DOMAIN_NAME environment variable, which
configures the domain name used in absolute URLs returned by the
application, for example in emails or Cargo responses.

The scenario that prompted this change is when an user creates an
account on the staging instance: before this commit the link they'd get
in the verification email was invalid, as the production application
doesn't know staging's email verification codes.

If the variable is omitted the domain will default to "crates.io".