categories: Add os::linux-apis
Linux has a large API surface that isn't shared with other
members of the Unix/Unix-like family, so having a separate
category for it allows crate authors to better distinguish
when writing bindings for Linux-specific interfaces.

This is consistent with the availability of `os::macos-apis`,
as macOS is another member of the Unix/Unix-like family with
its own large and discrete API surface.