add COM0 and LPT0 to the list of reserved crate names
In #695 we added all the documented reserved Windows file names to the
list of reserved crate names, to prevent people from registering crates
named that way and breaking all Windows users in the process.

Unfortunately it turns out that COM0 and LPT0 are *also* reserved on
Windows systems, but they're not documented in Microsoft's docs. This PR
adds both names to the reserved crate names list.

Note that the migration here is a no-op on the production crates.io
instance, as both names were manually added to the database before
opening this PR.