Log a flag for downloads with non-canonical crate names
This change will allow us to collect some data to see if it would be
practical to blindly redirect the request if there is an issue obtaining
a database connection from the pool.

We could instead include the corrected name, but that doesn't seem very
useful and `bot=` is already something I search the logs for
occasionally.