Reject requests with large page offset for blocked user-agents
`WEB_PAGE_OFFSET_UA_BLOCKLIST` is a comma seperated list of user-agents
that will be blocked if `WEB_MAX_ALLOWED_PAGE_OFFSET` is exceeded.

`WEB_MAX_ALLOWED_PAGE_OFFSET` defaults to 200.
`WEB_PAGE_OFFSET_UA_BLOCKLIST` defaults to the empty list. Because
empty user-agents are already blocked for non-download endpoints, the
check is disabled if this variable is empty or undefined.