Leverage readme_path to support relative paths from readme's not at root
Added unit tests to confirm things work.
Manually tested end to end and confirmed that the rendered
page has a properly working relative link as long as the Cargo.toml
is in the root of the repo.

We'll need to update cargo as mentioned in #3484 in order to
support Cargo.toml that aren't in the root of the repo