Use `minijinja` instead of `handlebars`
`minijinja` is explicitly built to be a light-weight templating system with minimal dependencies. Since we don't use a lot of the complexity that a full templating system like `handlebars` brings, we can reduce our number of transitive dependencies by replacing `handlebars` with `minijinja`. Note that it currently does not reduce the number of dependencies significantly, because other dependencies are still depending on these other transitive dependencies... 🙈