admin/render_readmes.rs to correctly identify the readme_path
Previously, it would try to get a field called `readme_file`
which never existed in the Cargo.toml manifest format. This
fixes it.

https://doc.rust-lang.org/cargo/reference/manifest.html#the-readme-field

I believe the confusion arose because the `cargo package` command
actually converts the readme contents -> readme field and the
Cargo.toml readme field -> `readme_file` field when making the
request to crates.io.

However, in the admin render_readme backfill, it is opening the
original Cargo.toml which has a `readme` field for the path.

I manually tested with a local backend/frontend and confirmed that
readme paths were being re-rendered correctly.